pyproject.toml
[project]
name = "assemblyai-tool"
version = "0.1.0"
description = "Simple CLI wrapper around the AssemblyAI Python SDK"
readme = "README.md"
authors = [
    { name = "Yaniv Aknin", email = "yaniv@aknin.name" }
]
requires-python = ">=3.12"
dependencies = [
    "assemblyai>=0.36.0",
    "python-dotenv>=1.0.1",
    "typer>=0.15.1",
]

[project.scripts]
aait = "assemblyai_tool:app"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
